I changed the setup of the database connection a bit. I created a central object DB in the database package and this is where ALL methods that query or update the database are called.
The way it is set up, this DB object assigns tasks to the
appropriate class (i.e. Query or Update) depending on the task at hand.
This way, the UI designer's job is easier since you just go directly
through DB. I figured this would be the most professional way to design
it (and easiest way to keep track of things).

In summary, if you want to get information from (or put information
in) the database:

1. define the method in the DB class
2. add a brief description of what you expect the method to do / fetch
3. include a TODO tag so I see it. I'll do the rest.